refactor: split out LLM-related code
