+++ src/raphael_backend_flask/process.py
@@ -1,14 +1,8 @@
 import json
-from typing import Any, Iterable
 
 import requests
 
-from health_misinfo_shared.fine_tuning import infer_claims
-from raphael_backend_flask.db import (
-    create_claim_extraction_run,
-    execute_sql,
-    update_claim_extraction_run,
-)
+from raphael_backend_flask.db import create_claim_extraction_run
 from raphael_backend_flask.youtube import download_captions, extract_title
 
 
@@ -23,12 +17,13 @@ def download_transcript(user_id: int, youtube_id: str) -> int:
     transcript = download_captions(video_html)
 
     # Add transcript text
-    return create_claim_extraction_run(
+    claim_extraction_run_id = create_claim_extraction_run(
         user_id,
         youtube_id,
         json.dumps(metadata),
         json.dumps(transcript),
     )
+    return claim_extraction_run_id
 
 
 def refine_offsets(claim: dict, transcript: dict) -> dict:
@@ -67,37 +62,3 @@ def refine_offsets(claim: dict, transcript: dict) -> dict:
     if end_idx < len(transcript):
         claim["offset_end_s"] = transcript[end_idx]["start"]
     return claim
-
-
-def extract_claims(run: dict) -> Iterable[dict[str, Any]]:
-    sentences = run["transcript"]
-    inferred_claims = infer_claims(run["id"], sentences)
-
-    for response in inferred_claims:
-        claims = response.get("response")
-        chunk = response.get("chunk")
-        for claim in claims:
-            labels_dict = claim.get("labels", {})
-            checkworthiness = claim.get("labels", {}).get("summary", "na")
-            # checkworthiness will be one of "worth checking", "may be worth checking" or "not worth checking"
-            parsed_claim = {
-                "run_id": run["id"],
-                "claim": claim["claim"],
-                "raw_sentence_text": claim["original_text"],
-                "labels": json.dumps(labels_dict),
-                "offset_start_s": chunk["start_offset"],
-            }
-            if chunk["end_offset"] is not None:
-                parsed_claim["offset_end_s"] = chunk["end_offset"]
-
-            parsed_claim = refine_offsets(parsed_claim, sentences)
-
-            execute_sql(
-                f"INSERT INTO inferred_claims ({', '.join(parsed_claim.keys())}) VALUES ({', '.join(['?'] * len(parsed_claim))})",
-                tuple(parsed_claim.values()),
-            )
-            parsed_claim["labels"] = labels_dict
-            yield parsed_claim
-
-    # Mark transcript done
-    update_claim_extraction_run(run["id"], status="complete")
